Tournament-reactor/Tournament-user-core/src/main/java/fr/webant/tournament/user/core/encryption/PasswordEncryptionUtil.java
package fr.webant.tournament.user.core.encryption;

import java.io.UnsupportedEncodingException;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;

/**
 * @author anthony
 */
public class PasswordEncryptionUtil {

    private static final String ALPHA_NUMERIC_STRING = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
    private static final Integer SALT_SIZE = 20;

    /**
     * Create a random salt with SALT_SIZE character(s). The salt contains characters from ALPHA_NUMERIC_STRING
     * @return a random salt
     */
    public static String generateRandomSalt() {
        int count = SALT_SIZE;
        StringBuilder builder = new StringBuilder();
        while (count-- != 0) {
            int character = (int) (Math.random() * ALPHA_NUMERIC_STRING.length());
            builder.append(ALPHA_NUMERIC_STRING.charAt(character));
        }
        return builder.toString();
    }

    /**
     * Hash a password with SHA-512 encryption
     * This method generate the same password that external tools with the same salt
     * @param passwordToHash uncrypted password which need to be hashed
     * @param salt the salt used
     * @return the hashed password
     */
    public static String hashPassword(String passwordToHash, String salt) {
        String generatedPassword = null;
        try {
            MessageDigest md = MessageDigest.getInstance("SHA-512");
            md.update(salt.getBytes("UTF-8"));
            byte[] bytes = md.digest(passwordToHash.getBytes("UTF-8"));
            StringBuilder sb = new StringBuilder();
            for (int i = 0; i < bytes.length; i++) {
                sb.append(Integer.toString((bytes[i] & 0xff) + 0x100, 16).substring(1));
            }
            generatedPassword = sb.toString();
        } catch (UnsupportedEncodingException | NoSuchAlgorithmException e) {
            e.printStackTrace();
        }
        return generatedPassword;
    }
}
